This is a different version of the project- some stuff-just trying to get it to work with Spring
