+++ .gitignore
@@ -31,4 +31,3 @@ build/
 
 ### VS Code ###
 .vscode/
-/bin/
