--- TeamFour.eml
@@ -1,101 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<component LANGUAGE_LEVEL="JDK_11" inheritJdk="true">
-	<output-test url="file://$MODULE_DIR$/target/test-classes"/>
-	<contentEntry url="file://$MODULE_DIR$">
-		<testFolder url="file://$MODULE_DIR$/src/test/java"/>
-		<excludeFolder url="file://$MODULE_DIR$/target"/>
-	</contentEntry>
-	<lib name="Maven: mysql:mysql-connector-java:8.0.26" scope="RUNTIME"/>
-	<lib name="Maven: org.springframework.boot:spring-boot-starter-test:2.5.5" scope="TEST"/>
-	<lib name="Maven: org.springframework.boot:spring-boot-test:2.5.5" scope="TEST"/>
-	<lib name="Maven: org.springframework.boot:spring-boot-test-autoconfigure:2.5.5" scope="TEST"/>
-	<lib name="Maven: com.jayway.jsonpath:json-path:2.5.0" scope="TEST"/>
-	<lib name="Maven: net.minidev:json-smart:2.4.7" scope="TEST"/>
-	<lib name="Maven: net.minidev:accessors-smart:2.4.7" scope="TEST"/>
-	<lib name="Maven: org.ow2.asm:asm:9.1" scope="TEST"/>
-	<lib name="Maven: jakarta.xml.bind:jakarta.xml.bind-api:2.3.3" scope="TEST"/>
-	<lib name="Maven: jakarta.activation:jakarta.activation-api:1.2.2" scope="TEST"/>
-	<lib name="Maven: org.assertj:assertj-core:3.19.0" scope="TEST"/>
-	<lib name="Maven: org.hamcrest:hamcrest:2.2" scope="TEST"/>
-	<lib name="Maven: org.junit.jupiter:junit-jupiter:5.7.2" scope="TEST"/>
-	<lib name="Maven: org.junit.jupiter:junit-jupiter-api:5.7.2" scope="TEST"/>
-	<lib name="Maven: org.apiguardian:apiguardian-api:1.1.0" scope="TEST"/>
-	<lib name="Maven: org.opentest4j:opentest4j:1.2.0" scope="TEST"/>
-	<lib name="Maven: org.junit.platform:junit-platform-commons:1.7.2" scope="TEST"/>
-	<lib name="Maven: org.junit.jupiter:junit-jupiter-params:5.7.2" scope="TEST"/>
-	<lib name="Maven: org.junit.jupiter:junit-jupiter-engine:5.7.2" scope="TEST"/>
-	<lib name="Maven: org.junit.platform:junit-platform-engine:1.7.2" scope="TEST"/>
-	<lib name="Maven: org.mockito:mockito-core:3.9.0" scope="TEST"/>
-	<lib name="Maven: net.bytebuddy:byte-buddy:1.10.22" scope="TEST"/>
-	<lib name="Maven: net.bytebuddy:byte-buddy-agent:1.10.22" scope="TEST"/>
-	<lib name="Maven: org.objenesis:objenesis:3.2" scope="TEST"/>
-	<lib name="Maven: org.mockito:mockito-junit-jupiter:3.9.0" scope="TEST"/>
-	<lib name="Maven: org.skyscreamer:jsonassert:1.5.0" scope="TEST"/>
-	<lib name="Maven: com.vaadin.external.google:android-json:0.0.20131108.vaadin1" scope="TEST"/>
-	<lib name="Maven: org.springframework:spring-test:5.3.10" scope="TEST"/>
-	<lib name="Maven: org.xmlunit:xmlunit-core:2.8.2" scope="TEST"/>
-	<levels>
-		<level name="Maven: org.springframework.boot:spring-boot-starter-web:2.5.5" value="project"/>
-		<level name="Maven: org.springframework.boot:spring-boot-starter:2.5.5" value="project"/>
-		<level name="Maven: org.springframework.boot:spring-boot:2.5.5" value="project"/>
-		<level name="Maven: org.springframework.boot:spring-boot-autoconfigure:2.5.5" value="project"/>
-		<level name="Maven: org.springframework.boot:spring-boot-starter-logging:2.5.5" value="project"/>
-		<level name="Maven: ch.qos.logback:logback-classic:1.2.6" value="project"/>
-		<level name="Maven: ch.qos.logback:logback-core:1.2.6" value="project"/>
-		<level name="Maven: org.apache.logging.log4j:log4j-to-slf4j:2.14.1" value="project"/>
-		<level name="Maven: org.apache.logging.log4j:log4j-api:2.14.1" value="project"/>
-		<level name="Maven: org.slf4j:jul-to-slf4j:1.7.32" value="project"/>
-		<level name="Maven: jakarta.annotation:jakarta.annotation-api:1.3.5" value="project"/>
-		<level name="Maven: org.yaml:snakeyaml:1.28" value="project"/>
-		<level name="Maven: org.springframework.boot:spring-boot-starter-json:2.5.5" value="project"/>
-		<level name="Maven: com.fasterxml.jackson.core:jackson-databind:2.12.5" value="project"/>
-		<level name="Maven: com.fasterxml.jackson.core:jackson-annotations:2.12.5" value="project"/>
-		<level name="Maven: com.fasterxml.jackson.core:jackson-core:2.12.5" value="project"/>
-		<level name="Maven: com.fasterxml.jackson.datatype:jackson-datatype-jdk8:2.12.5" value="project"/>
-		<level name="Maven: com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.12.5" value="project"/>
-		<level name="Maven: com.fasterxml.jackson.module:jackson-module-parameter-names:2.12.5" value="project"/>
-		<level name="Maven: org.springframework.boot:spring-boot-starter-tomcat:2.5.5" value="project"/>
-		<level name="Maven: org.apache.tomcat.embed:tomcat-embed-core:9.0.53" value="project"/>
-		<level name="Maven: org.apache.tomcat.embed:tomcat-embed-el:9.0.53" value="project"/>
-		<level name="Maven: org.apache.tomcat.embed:tomcat-embed-websocket:9.0.53" value="project"/>
-		<level name="Maven: org.springframework:spring-web:5.3.10" value="project"/>
-		<level name="Maven: org.springframework:spring-beans:5.3.10" value="project"/>
-		<level name="Maven: org.springframework:spring-webmvc:5.3.10" value="project"/>
-		<level name="Maven: org.springframework:spring-aop:5.3.10" value="project"/>
-		<level name="Maven: org.springframework:spring-context:5.3.10" value="project"/>
-		<level name="Maven: org.springframework:spring-expression:5.3.10" value="project"/>
-		<level name="Maven: mysql:mysql-connector-java:8.0.26" value="project"/>
-		<level name="Maven: org.projectlombok:lombok:1.18.20" value="project"/>
-		<level name="Maven: org.springframework.boot:spring-boot-starter-test:2.5.5" value="project"/>
-		<level name="Maven: org.springframework.boot:spring-boot-test:2.5.5" value="project"/>
-		<level name="Maven: org.springframework.boot:spring-boot-test-autoconfigure:2.5.5" value="project"/>
-		<level name="Maven: com.jayway.jsonpath:json-path:2.5.0" value="project"/>
-		<level name="Maven: net.minidev:json-smart:2.4.7" value="project"/>
-		<level name="Maven: net.minidev:accessors-smart:2.4.7" value="project"/>
-		<level name="Maven: org.ow2.asm:asm:9.1" value="project"/>
-		<level name="Maven: org.slf4j:slf4j-api:1.7.32" value="project"/>
-		<level name="Maven: jakarta.xml.bind:jakarta.xml.bind-api:2.3.3" value="project"/>
-		<level name="Maven: jakarta.activation:jakarta.activation-api:1.2.2" value="project"/>
-		<level name="Maven: org.assertj:assertj-core:3.19.0" value="project"/>
-		<level name="Maven: org.hamcrest:hamcrest:2.2" value="project"/>
-		<level name="Maven: org.junit.jupiter:junit-jupiter:5.7.2" value="project"/>
-		<level name="Maven: org.junit.jupiter:junit-jupiter-api:5.7.2" value="project"/>
-		<level name="Maven: org.apiguardian:apiguardian-api:1.1.0" value="project"/>
-		<level name="Maven: org.opentest4j:opentest4j:1.2.0" value="project"/>
-		<level name="Maven: org.junit.platform:junit-platform-commons:1.7.2" value="project"/>
-		<level name="Maven: org.junit.jupiter:junit-jupiter-params:5.7.2" value="project"/>
-		<level name="Maven: org.junit.jupiter:junit-jupiter-engine:5.7.2" value="project"/>
-		<level name="Maven: org.junit.platform:junit-platform-engine:1.7.2" value="project"/>
-		<level name="Maven: org.mockito:mockito-core:3.9.0" value="project"/>
-		<level name="Maven: net.bytebuddy:byte-buddy:1.10.22" value="project"/>
-		<level name="Maven: net.bytebuddy:byte-buddy-agent:1.10.22" value="project"/>
-		<level name="Maven: org.objenesis:objenesis:3.2" value="project"/>
-		<level name="Maven: org.mockito:mockito-junit-jupiter:3.9.0" value="project"/>
-		<level name="Maven: org.skyscreamer:jsonassert:1.5.0" value="project"/>
-		<level name="Maven: com.vaadin.external.google:android-json:0.0.20131108.vaadin1" value="project"/>
-		<level name="Maven: org.springframework:spring-core:5.3.10" value="project"/>
-		<level name="Maven: org.springframework:spring-jcl:5.3.10" value="project"/>
-		<level name="Maven: org.springframework:spring-test:5.3.10" value="project"/>
-		<level name="Maven: org.xmlunit:xmlunit-core:2.8.2" value="project"/>
-	</levels>
-</component>
