--- src/main/java/application/events/TimerStart.java
@@ -1,24 +1,0 @@
-package application.events;
-
-public class TimerStart {
-	
-    private static TimerStart instance;
-
-    /**
-     * This is a singleton class. Hence the private constructor.
-     */
-
-    private TimerStart() {}
-
-    /**
-     * Returns the only instance of the class.
-     *
-     * @return the only instance
-     */
-
-    public static TimerStart instance() {
-	if (instance == null) instance = new TimerStart();
-	return instance;
-    }
-
-}
