--- src/main/java/application/events/TimerStop.java
@@ -1,24 +1,0 @@
-package application.events;
-
-public class TimerStop {
-	
-    private static TimerStop instance;
-
-    /**
-     * This is a singleton class. Hence the private constructor.
-     */
-
-    private TimerStop() {}
-
-    /**
-     * Returns the only instance of the class.
-     *
-     * @return the only instance
-     */
-
-    public static TimerStop instance() {
-	if (instance == null) instance = new TimerStop();
-	return instance;
-    }
-
-}
