+++ src/main/java/controller/GameController.java
@@ -1,0 +1,8 @@
+package controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class GameController {
+
+}
