+++ src/main/java/entity/Point.java
@@ -1,0 +1,23 @@
+package entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+
+@Data
+@Entity
+@NoArgsConstructor
+@AllArgsConstructor
+public class Point {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    public long pointID;
+    @Column(nullable = false)
+    private int x;
+    private int y;
+
+}
