+++ src/main/java/tech/dunny/gameproject/repos/PointRepository.java
@@ -1,0 +1,8 @@
+package tech.dunny.gameproject.repos;
+
+import entity.Point;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+public interface PointRepository extends JpaRepository<Point, Long> {
+}
