+++ src/main/resources/application.properties
@@ -1,1 +1,8 @@
+spring.jpa.show-sql=true
+spring.datasource.url=jdbc:mysql://192.168.117.45:3306/gamedb
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.username=root
+spring.datasource.password=root
+spring.jpa.hibernate.ddl-auto=update
 
+spring.jpa.open-in-view=false
