--- src/main/resources/static/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Class Project</title>
-</head>
-<body>
-    <H1>This is where our game will be</H1>
-    <H2>This is for ICS 499 - it will be the place where you can play <br>our game.</H2>
-</body>
-</html>
