--- src/test/java/junit/TestSuite.java
@@ -1,18 +1,0 @@
-package junit;
-
-import static org.junit.jupiter.api.Assertions.*;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TestSuite {
-    @Test
-    void contextLoads() {
-	fail("Not yet implemented");
-    }
-    
-    @Test
-    void otherTests() {
-	fail("Not yet implemented");
-    }
-}
