+++ src/test/java/tech/dunny/gameproject/GameProjectApplicationTests.java
@@ -1,0 +1,13 @@
+package tech.dunny.gameproject;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GameProjectApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
